move rspec-only support file under spec/support
